# documentation/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HW-9",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "variables": {}
    }
  ],
  "paths": {
    "/user/register": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "User Create Register",
        "operationId": "user/register",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UsersCreateRequest"
                  },
                  {
                    "example": {
                      "id": 102,
                      "email": "agis@email.com",
                      "gender": "Female",
                      "password": "qwerty11",
                      "role": "Constructor Worker"
                    }
                  }
                ]
              },
              "example": {
                "id": 102,
                "email": "agis@email.com",
                "gender": "Female",
                "password": "qwerty11",
                "role": "Constructor Worker"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/users": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Login",
        "operationId": "LoginUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LoginUsersRequest"
                  },
                  {
                    "example": {
                      "email": "jborsnall8@mediafire.com",
                      "password": "rFEoyMIhq3mF"
                    }
                  }
                ]
              },
              "example": {
                "email": "jborsnall8@mediafire.com",
                "password": "rFEoyMIhq3mF"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/movies": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "Movies",
        "operationId": "Movies",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Movies Create",
        "operationId": "MoviesCreate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/MoviesCreateRequest"
                  },
                  {
                    "example": {
                      "id": 104,
                      "title": "a Taxi",
                      "genres": "Action",
                      "year": 2023
                    }
                  }
                ]
              },
              "example": {
                "id": 104,
                "title": "a Taxi",
                "genres": "Action",
                "year": 2023
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/movies/103": {
      "put": {
        "tags": [
          "Misc"
        ],
        "summary": "Movies Update",
        "operationId": "MoviesUpdate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/MoviesUpdateRequest"
                  },
                  {
                    "example": {
                      "id": 103,
                      "title": "Blade X",
                      "genres": "Action|Crime",
                      "year": 2023
                    }
                  }
                ]
              },
              "example": {
                "id": 103,
                "title": "Blade X",
                "genres": "Action|Crime",
                "year": 2023
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/movies/104": {
      "delete": {
        "tags": [
          "Misc"
        ],
        "summary": "Movies Delete",
        "operationId": "MoviesDelete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    }
  },
  "components": {
    "schemas": {
      "LoginUsersRequest": {
        "title": "LoginUsersRequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "jborsnall8@mediafire.com",
          "password": "rFEoyMIhq3mF"
        }
      },
      "LoginAdminRequest": {
        "title": "LoginAdminRequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "oainger0@craigslist.org",
          "password": "KcAk6Mrg7DRM"
        }
      },
      "MoviesCreateRequest": {
        "title": "MoviesCreateRequest",
        "required": [
          "id",
          "title",
          "genres",
          "year"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "genres": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id": 104,
          "title": "a Taxi",
          "genres": "Action",
          "year": 2023
        }
      },
      "MoviesUpdateRequest": {
        "title": "MoviesUpdateRequest",
        "required": [
          "id",
          "title",
          "genres",
          "year"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "genres": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id": 103,
          "title": "Blade X",
          "genres": "Action|Crime",
          "year": 2023
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
  "tags": [
    {
      "name": "Misc",
      "description": ""
    }
  ]
}